% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangements.R
\name{k_combinations}
\alias{k_combinations}
\alias{all_combinations}
\title{Combinations in draws of size k}
\usage{
k_combinations(v, k = NULL)

all_combinations(v)
}
\arguments{
\item{v}{a vector to be drawn}

\item{k}{an integer, the number of items drawn}
}
\description{
Wraps \code{\link[arrangements:combinations]{arrangements::combinations()}} for use with \code{\link[=groups_split]{groups_split()}}.
\itemize{
\item \code{\link[=k_combinations]{k_combinations()}} draws combinations of size k
\item \code{\link[=all_combinations]{all_combinations()}} draws combinations of size 1 to k
}
}
\examples{
v <- LETTERS[1:3]

k_combinations(v)

k_combinations(v, 1)

k_combinations(v, 2)

all_combinations(v)
}
