% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangements.R
\name{individuals}
\alias{individuals}
\title{Individual items from a vector}
\usage{
individuals(v)
}
\arguments{
\item{v}{a vector to be drawn}
}
\description{
For use with \code{\link[=groups_split]{groups_split()}}.  Shorthand for \code{k_combinations(v, k = 1)}.
Longhand for \code{as.list(v)}.
}
\examples{
v <- LETTERS[1:3]

individuals(v)
}
