% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangements.R
\name{k_permutations}
\alias{k_permutations}
\alias{all_permutations}
\title{Permutations in draws of size k}
\usage{
k_permutations(v, k = NULL)

all_permutations(v)
}
\arguments{
\item{v}{a vector to be drawn}

\item{k}{an integer, the number of items drawn}
}
\description{
Wraps \code{\link[arrangements:permutations]{arrangements::permutations()}} for use with \code{\link[=groups_split]{groups_split()}}.
\itemize{
\item \code{\link[=k_permutations]{k_permutations()}} draws permutations of size k
\item \code{\link[=all_permutations]{all_permutations()}} draws permutations of size 1 to k
}

This is only useful when later using \code{dplyr::arrange()} with \code{.by_group = TRUE}, otherwise permutations of the same combination are equivalent.
}
\examples{
v <- LETTERS[1:3]

k_permutations(v)

k_permutations(v, 1)

k_permutations(v, 2)

all_permutations(v)
}
