% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupie-package.R
\docType{package}
\name{groupie-package}
\alias{groupie}
\alias{groupie-package}
\title{groupie: Map Functions to Grouped Data Frames}
\description{
The groupie package provides `groups_split()` to create copies of a
    data frame with different combinations of grouping variables.  This is
    useful for summarising data several ways in one go.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nacnudus/groupie}
  \item Report bugs at \url{https://github.com/nacnudus/groupie/issues}
}

}
\author{
\strong{Maintainer}: Duncan Garmonsway \email{nacnudus@gmail.com}

}
\keyword{internal}
